requirements.txt
pillow==9.3.0
cryptography==39.0.1
typer
requests
openai